Saving changes for the next task

+++ GymCRM/src/main/java/com/haykz/GymCRM/model/Training.java
@@ -32,5 +32,5 @@ public class Training {
     private Date trainingDate;
 
     @Column(nullable = false)
-    private BigDecimal trainingDuration;
+    private Integer trainingDuration;
 }
